requirements.txt
diffusers==0.9.0
numpy==1.19.2
openai==0.26.1
Pillow==9.4.0
RealESRGAN==0.3.0
requests==2.28.1
stability_sdk==0.3.1
torch==1.13.1
torchvision==0.14.1
transformers==4.24.0
